pyproject.toml
[project]
name = "tripartite-rooms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.39.0",
    "pydantic>=2.10.1",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]
